+++ server/apis/videoRoutes.js
@@ -2,18 +2,19 @@ const express=require('express');
 const router=express.Router();
 const FitnessVideo=require('../models/videoModel')
 const imagekit=require('../utils/imageKitCredentials')
+const { deleteImageKitFiles } = require('../utils/imageKitCleanup');
 const mongoose=require('mongoose');
 const User=require('../models/User');
 const { required } = require('joi');
 const multer=require('multer');
 const upload=multer({ dest:'files/'})
 const  fs = require('fs');
-const { isVideoAuthor, isLoggedIn, isNotVideoAuthor, isNotCoach } = require('../middleware');
+const { isVideoAuthor, isLoggedIn, isNotVideoAuthor, isNotCoach, isCoach } = require('../middleware');
 const Program = require('../models/Program');
 
 router.get('/allvideos' , async(req,res)=>{
     try{
-        let fitnessVideos=await FitnessVideo.find({}).limit(3).populate('coach');
+        let fitnessVideos=await FitnessVideo.find({}).populate('coach');
         // let coach=fitnessVideos
         // let coach=await User.find({});
         console.log("This is fitness videos", fitnessVideos);
@@ -36,7 +37,7 @@ router.get('/show/:data', async(req,res)=>{
     }
 })
 
-router.delete('/delete/:id',isLoggedIn, isVideoAuthor ,async (req,res)=>{
+router.delete('/delete/:id',isLoggedIn, isCoach, isVideoAuthor ,async (req,res)=>{
     try {
         let id=req.params.id;
         let programs=await Program.find({});
@@ -49,16 +50,14 @@ router.delete('/delete/:id',isLoggedIn, isVideoAuthor ,async (req,res)=>{
                 }
             }
         }
-        let fitnessVideos=await FitnessVideo.findById(id)
-        imagekit.deleteFile(fitnessVideos.fileId, function(error, result) {
-            if(error) console.log(error);
-            else console.log(result);
+        let fitnessVideos=await FitnessVideo.findById(id);
+        if (!fitnessVideos) {
+            return res.status(404).json({ msg: "Video not found" });
+        }
+        await deleteImageKitFiles({
+            fileIds: [fitnessVideos.fileId, fitnessVideos.imgFileId],
         });
-        imagekit.deleteFile(fitnessVideos.imgFileId, function(error, result) {
-            if(error) console.log(error);
-            else console.log(result);
-        });
-        let response=await FitnessVideo.findByIdAndDelete(id);
+        await FitnessVideo.findByIdAndDelete(id);
         res.status(201).json({msg: "Gotcha" });
     } catch (e) {
         res.status(400).json({msg: "Something went wrong..." });
@@ -106,19 +105,47 @@ router.delete('/delete/:id',isLoggedIn, isVideoAuthor ,async (req,res)=>{
 //     }
 // }
 
-router.patch('/edit/:id',isLoggedIn,isVideoAuthor  , upload.none(), async(req,res)=>{
-    try{
-        let {name,tags,id}=req.body;
-        let arr=tags.split(",");
-        let found=await FitnessVideo.findById(id);
-        let {fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl}=found;
-        await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});
-        res.status(201).json({msg: "Gotcha"} );
-    }catch(e){
-        res.status(400).json({msg: "Something went wrong..." });
-    }
+router.patch('/edit/f', isLoggedIn, isCoach, upload.fields([
+    { name: 'file', maxCount: 1 },
+]), isVideoAuthor, async(req,res)=>{
+  
+        fs.readFile(req.files.file[0].path, function(err, data) {
+            if (err) throw err; // Fail if the file can't be read.
+            else{
+                var {name,tags,id}=req.body;
+                imagekit.upload({
+                file : data, //required
+                fileName : `${String(name).trim().replace(/\s+/g, "_")}.mp4`, //required
+                folder: '/fitness'
+                }, async function(error, result) {
+                if(error) console.log(error);
+                else{
+                        var {name,tags,id}=req.body;
+                        let arr=tags.split(",");
+                        let found=await FitnessVideo.findById(id);
+                        let {fileId, imgFilePath , imgFileUrl, imgFileId}=found;
+                        const oldVideoFileId = fileId;
+                        fileId=result.fileId;
+                        await deleteImageKitFiles({ fileIds: [oldVideoFileId] });
+                        let filePath=result.filePath;
+                        let fileUrl=result.url;
+                        try {
+                            
+                            
+                            await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});
+                            res.status(201).json({msg: "Gotcha"} );
+                        } catch (e) {
+                            res.status(400).json({msg: "Something went wrong..."});
+                        }
+                }
+                });
+            }
+          });
+        
+   
 })
-router.patch('/edit/imgf/:id',isLoggedIn,isVideoAuthor  ,upload.fields([
+
+router.patch('/edit/imgf/:id',isLoggedIn, isCoach, isVideoAuthor  ,upload.fields([
     { name: 'imgFile', maxCount: 1 }
    ]), async(req,res)=>{
   
@@ -137,15 +164,13 @@ router.patch('/edit/imgf/:id',isLoggedIn,isVideoAuthor  ,upload.fields([
                         let arr=tags.split(",");
                         let found=await FitnessVideo.findById(id);
                         let {fileId, fileUrl, filePath, imgFileId}=found;
-                        imagekit.deleteFile(imgFileId, function(error, result) {
-                            if(error) console.log(error);
-                            else console.log(result);
-                        });
+                        const oldThumbnailFileId = imgFileId;
                         imgFileId=result.fileId;
+                        await deleteImageKitFiles({ fileIds: [oldThumbnailFileId] });
                         let imgFilePath=result.filePath;
                         let imgFileUrl=result.url;
                         try {
-                            await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});                        res.status(201).json({msg: "New video added"});
+                            await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});
                             res.status(201).json({msg: "Gotcha"} );
                         } catch (e) {
                             res.status(400).json({msg: "Something went wrong..."});
@@ -158,64 +183,17 @@ router.patch('/edit/imgf/:id',isLoggedIn,isVideoAuthor  ,upload.fields([
    
 })
 
-
-
-
-// router.patch('/edit/:id',isLoggedIn,isVideoAuthor  , upload.none(), async(req,res)=>{
-//     try{
-//         let {name,tags,id}=req.body;
-//         console.log("This  is name tag id", name,tags,id);
-//         let arr=tags.split(",");
-//         console.log("This  is name tag id", name,tags,id,arr);
-//         let found=await FitnessVideo.findById(id);
-//         console.log("this is found",found);
-//         let {fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl}=found;
-//         await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});
-//         res.status(201).json({msg: "Gotcha"} );
-//     }catch(e){
-//         res.status(400).json({msg: "Something went wrong..." });
-//     }
-// })
-router.patch('/edit/f',isLoggedIn , isVideoAuthor ,upload.fields([
-    { name: 'file', maxCount: 1 },
-   ]), async(req,res)=>{
-  
-        fs.readFile(req.files.file[0].path, function(err, data) {
-            if (err) throw err; // Fail if the file can't be read.
-            else{
-                var {name,tags,id}=req.body;
-                imagekit.upload({
-                file : data, //required
-                fileName : name, //required
-                folder: '/fitness'
-                }, async function(error, result) {
-                if(error) console.log(error);
-                else{
-                        var {name,tags,id}=req.body;
-                        let arr=tags.split(",");
-                        let found=await FitnessVideo.findById(id);
-                        let {fileId, imgFilePath , imgFileUrl, imgFileId}=found;
-                        imagekit.deleteFile(fileId, function(error, result) {
-                            if(error) console.log(error);
-                            else console.log(result);
-                        });
-                        fileId=result.fileId;
-                        let filePath=result.filePath;
-                        let fileUrl=result.url;
-                        try {
-                            
-                            
-                            await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});                        res.status(201).json({msg: "New video added"});
-                            res.status(201).json({msg: "Gotcha"} );
-                        } catch (e) {
-                            res.status(400).json({msg: "Something went wrong..."});
-                        }
-                }
-                });
-            }
-          });
-        
-   
+router.patch('/edit/:id',isLoggedIn, isCoach, isVideoAuthor  , upload.none(), async(req,res)=>{
+    try{
+        let {name,tags,id}=req.body;
+        let arr=tags.split(",");
+        let found=await FitnessVideo.findById(id);
+        let {fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl}=found;
+        await FitnessVideo.findByIdAndUpdate(id,{name, tags:arr, fileId, imgFileId, filePath , imgFilePath, fileUrl, imgFileUrl});
+        res.status(201).json({msg: "Gotcha"} );
+    }catch(e){
+        res.status(400).json({msg: "Something went wrong..." });
+    }
 })
 
 router.get('/getall', isLoggedIn,async(req,res)=>{
@@ -262,7 +240,7 @@ router.post('/changevideolike',isLoggedIn, async(req,res)=>{
 router.get('/allvideos/:tag' , async(req,res)=>{
     try{
         let tag = req.params.tag;
-        let fitnessVideos = await FitnessVideo.find({ tags: tag }).limit(5).populate('coach');        // let coach=fitnessVideos
+        let fitnessVideos = await FitnessVideo.find({ tags: tag }).populate('coach');        // let coach=fitnessVideos
         // let coach=await User.find({});
 
         // fitnessVideos=await fitnessVideos.populate('coach');
@@ -282,10 +260,16 @@ router.get("/test", (req, res) => {
 router.patch("/addrating/:id", isLoggedIn, isNotVideoAuthor,isNotCoach, upload
 .none(), async (req, res) => {
   try { 
-      let user = req.user;
-      let id = req.params.id;
+    let user = req.user;
+    let id = req.params.id;
     let { userRating, newRating, newRatingCount } = req.body;
     let found = await FitnessVideo.findById(id);
+
+    for (let i = 0; i < found.rating.length; i++) {
+      if (found.rating[i].user.toString() === user._id.toString()) {
+        return res.status(400).json({ msg: "You have already rated this video" });
+      }
+    }
 
     await FitnessVideo.findByIdAndUpdate(id, {
       currentRating: newRating,
@@ -297,4 +281,26 @@ router.patch("/addrating/:id", isLoggedIn, isNotVideoAuthor,isNotCoach, upload
     res.status(400).json({ msg: "Something went wrong..." });
   }
 });
+router.post('/video/addreview/:id', isLoggedIn, isNotVideoAuthor, async (req, res) => {
+  try {
+    const userId = req.user._id;
+    const id = req.params.id;
+    let { review } = req.body;
+    review = review.trim();
+    if (review.length === 0) {
+      return res.status(400).json({ msg: "Review cannot be empty" });
+    }
+    let found = await FitnessVideo.findById(id);
+    if (!found) {
+      return res.status(404).json({ msg: "Video not found" });
+    }
+    found.reviews.push({ review, user: userId });
+    await found.save();
+    res.status(201).json({ msg: "Gotcha", data: found });
+  } catch (e) {
+    console.error(e);
+    res.status(400).json({ msg: "Something went wrong..." });
+  }
+});
+
 module.exports=router;
